Seed the psueod-random number generators
"The GPSS/360 Random Number Generator" by Harry Felder does not give
enough information on the algorithm (most unfortunately, it says that
there is an 8x1 array of base numbers, but doesn't say what they are).
He references the GPSS/360 User's Manual from IBM, which I could not
find anywhere online except for the Computer History Museum. The CHM had
the First Edition of the manual, but Felder references what appears to
be the Third Edition (Publication Number ends in "-2", CHM's end in
"-0"). I didn't find any description of the RNG algorithm in the CHM's
copy of the User's Manual, so I presume it was only added after the
First Edition.

Without the original GPSS's pseudo-random number generation algorithm, I
can still try to do a similar thing. The random number generators in
gpss.py will use Python's built-in `random` module and use the seed of
134630. This seed yields results similar to the results provided in "A
GPSS Primer" by Thomas J. Schriber for a few example simulation
programs. However, that fact isn't important; the fact that the random
numbers generated are reproducable is.