Update project description
Despite the lack of numerous features of GPSS, I'd say gpss.py is now
decent--it should be usable for simple simulations, so the "no-good"
part should go.

I also feel that "silly" and "dumb" no longer belong either because I've
been taking this project fairly seriously and I aim to make it at least
somewhat as featureful as GPSS. Yes, recreating a now almost 60-year-old
program might be silly and dumb, but I'm finding it really interesting
to see how one would create, run, and analyze computer simulations in
the 60s and 70s!