Fix incorrect comment discardment
Previously, the parser would attempt to remove a semicolon-style comment
first, before even checking for an asterisk-style comment. This would
wrongly read '*' as a regular field in the following example:

    SEIZE Cookie *; Oops!

The parser would find the semicolon and discard the characters including
and following the semicolon, leaving "SEIZE Cookie *" to parse as a
statement. (A semicolon followed by an asterisk would work fine, since
the semicolon is found first anyway.)

Instead, check for a semicolon AND asterisk, removing the characters
including and following whichever appears first.